refactor datadescriptor encryption function
